add file like tests

+++ application/files/file_management.js
@@ -124,13 +124,29 @@ exports.editFile = async(ctx, next) => {
 }
 
 exports.deleteFile = async(ctx, next) => {
+
     let body = ctx.request.body;
+    let verified = await jwtChecker.decodeAuth(ctx);
+    // let verified = true;
 
-    let controller = new Controller();
-    let result = await controller.deleteFile(body);
-    ctx.body = result;
+    if (verified === false) {
+        let result = {
+            "status": 500,
+            "err_message": "authorization code invalid"
+        }
+        console.log("authorization code invalid");
+        ctx.body = result;
+        await next();
+        
+    } else {
+        body["email"] = verified;
+        let controller = new Controller();
+        let result = await controller.deleteFile(body);
+        ctx.body = result;
 
-    await next();
+        await next();
+    }
+
 }
 
 exports.likeFile = async(ctx, next) => {
